swh.auth.pytest_plugin: Make decoded_token consistent with user_info
Tester will want to craft carefully the keycloak mock so it returns
consistent data.

Currently the decoded_token and user_info are hard-coded to consistent values. So using
the default keycloak mock is ok. As soon as other testers will want another user_info,
the decoded token will diverge and not return the correct user_info subpart of the
decoded token.

This commits fixes such behavior and allows the decoded token to stay consistent with
the user_info.

Related to T2858