cli-auth-config: explicitly message about skipped update
At the end of a auth-config command run, we ask the user if the newly
obtained token should be written on disk or discarded.

The default is to discard it (which seems a bit weird, but is not the
point here). Unlike what we do on write, where a green confirmation
message is displayed, we do not display any message when discarding.
So a user doing things a bit too fast can imagine that the update went
all fine as it just display a wall of green messages.

So we add a simple yellow message pointing out that we ultimately did
not write the new token. That should fix the problem.